Add previous command
